Decrypt the secret signing key on all branches (except PRs)
Previously the secret keyring was only decrypted on the master branch (but without a check for the repository/fork that is built), not on all branches. But the Gradle build now requires the build to be signed on all branches of the repository `osmlab/josm-atlas` (see https://github.com/osmlab/josm-atlas/pull/13/files#r169728452).

Now the decryption only depends on the availability of the environment variables used for decryption ($encrypted_7407e5ec1bc3_key, encrypted_7407e5ec1bc3_iv, …).
These variables are not available for forks and pull requests. As soon as they'd be available to a malicious party, the encrypted key would otherwise be compromised.